Create-Office365-Mailbox
This script will create a new mailbox in Office 365 and configure settings that enable mailbox auditing, disable clutter, and more. Once the mailbox is created, it will connect to IT Glue and create a new contact at the appropriate customer organization, and then securely embeds the randomly generated mailbox password to the contact.